Move objectstoreprefetch over cacheprefetch nodes. Minor fixes.
